+++ adapters/adnuntius/adnuntius.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"net/http"
-	"net/url"
+
 	"strconv"
 	"strings"
 
@@ -17,80 +17,6 @@ import (
 	"github.com/prebid/prebid-server/v3/util/jsonutil"
 	"github.com/prebid/prebid-server/v3/util/timeutil"
 )
-
-type QueryString map[string]string
-type adapter struct {
-	time      timeutil.Time
-	endpoint  string
-	extraInfo string
-}
-type adnAdunit struct {
-	AuId       string    `json:"auId"`
-	TargetId   string    `json:"targetId"`
-	Dimensions [][]int64 `json:"dimensions,omitempty"`
-	MaxDeals   int       `json:"maxDeals,omitempty"`
-}
-
-type extDeviceAdnuntius struct {
-	NoCookies bool `json:"noCookies,omitempty"`
-}
-type siteExt struct {
-	Data interface{} `json:"data"`
-}
-
-type adnAdvertiser struct {
-	LegalName string `json:"legalName,omitempty"`
-	Name      string `json:"name,omitempty"`
-}
-
-type Ad struct {
-	Bid struct {
-		Amount   float64
-		Currency string
-	}
-	NetBid struct {
-		Amount float64
-	}
-	GrossBid struct {
-		Amount float64
-	}
-	DealID            string `json:"dealId,omitempty"`
-	AdId              string
-	CreativeWidth     string
-	CreativeHeight    string
-	CreativeId        string
-	LineItemId        string
-	Html              string
-	DestinationUrls   map[string]string
-	AdvertiserDomains []string
-	Advertiser        adnAdvertiser `json:"advertiser,omitempty"`
-}
-
-type AdUnit struct {
-	AuId       string
-	TargetId   string
-	Html       string
-	ResponseId string
-	Ads        []Ad
-	Deals      []Ad `json:"deals,omitempty"`
-}
-
-type AdnResponse struct {
-	AdUnits []AdUnit
-}
-type adnMetaData struct {
-	Usi string `json:"usi,omitempty"`
-}
-type adnRequest struct {
-	AdUnits   []adnAdunit `json:"adUnits"`
-	MetaData  adnMetaData `json:"metaData,omitempty"`
-	Context   string      `json:"context,omitempty"`
-	KeyValues interface{} `json:"kv,omitempty"`
-}
-
-type RequestExt struct {
-	Bidder adnAdunit `json:"bidder"`
-}
 
 const defaultNetwork = "default"
 const defaultSite = "unknown"
@@ -107,110 +33,13 @@ func Builder(bidderName openrtb_ext.BidderName, config config.Adapter, server co
 	return bidder, nil
 }
 
-func (a *adapter) MakeRequests(request *openrtb2.BidRequest, reqInfo *adapters.ExtraRequestInfo) ([]*adapters.RequestData, []error) {
-	return a.generateRequests(*request)
-}
-
-func setHeaders(ortbRequest openrtb2.BidRequest) http.Header {
-
-	headers := http.Header{}
-	headers.Add("Content-Type", "application/json;charset=utf-8")
-	headers.Add("Accept", "application/json")
-	if ortbRequest.Device != nil {
-		if ortbRequest.Device.IP != "" {
-			headers.Add("X-Forwarded-For", ortbRequest.Device.IP)
-		}
-		if ortbRequest.Device.UA != "" {
-			headers.Add("user-agent", ortbRequest.Device.UA)
-		}
-	}
-	return headers
-}
-
-func makeEndpointUrl(ortbRequest openrtb2.BidRequest, a *adapter, noCookies bool) (string, []error) {
-	uri, err := url.Parse(a.endpoint)
-	endpointUrl := a.endpoint
-	if err != nil {
-		return "", []error{fmt.Errorf("failed to parse Adnuntius endpoint: %v", err)}
-	}
-
-	gdpr, consent, err := getGDPR(&ortbRequest)
-	if err != nil {
-		return "", []error{fmt.Errorf("failed to parse Adnuntius endpoint: %v", err)}
-	}
-
-	if !noCookies {
-		var deviceExt extDeviceAdnuntius
-		if ortbRequest.Device != nil && ortbRequest.Device.Ext != nil {
-			if err := jsonutil.Unmarshal(ortbRequest.Device.Ext, &deviceExt); err != nil {
-				return "", []error{fmt.Errorf("failed to parse Adnuntius endpoint: %v", err)}
-			}
-		}
-
-		if deviceExt.NoCookies {
-			noCookies = true
-		}
-	}
-
-	_, offset := a.time.Now().Zone()
-	tzo := -offset / minutesInHour
-
-	q := uri.Query()
-	if gdpr != "" {
-		endpointUrl = a.extraInfo
-		q.Set("gdpr", gdpr)
-	}
-
-	if consent != "" {
-		q.Set("consentString", consent)
-	}
-
-	if noCookies {
-		q.Set("noCookies", "true")
-	}
-
-	q.Set("tzo", fmt.Sprint(tzo))
-	q.Set("format", "prebidServer")
-
-	url := endpointUrl + "?" + q.Encode()
-	return url, nil
-}
-
-func getImpSizes(imp openrtb2.Imp) [][]int64 {
-
-	if len(imp.Banner.Format) > 0 {
-		sizes := make([][]int64, len(imp.Banner.Format))
-		for i, format := range imp.Banner.Format {
-			sizes[i] = []int64{format.W, format.H}
-		}
-
-		return sizes
-	}
-
-	if imp.Banner.W != nil && imp.Banner.H != nil {
-		size := make([][]int64, 1)
-		size[0] = []int64{*imp.Banner.W, *imp.Banner.H}
-		return size
-	}
-
-	return nil
-}
-
-/*
-Generate the requests to Adnuntius to reduce the amount of requests going out.
-*/
 func (a *adapter) generateRequests(ortbRequest openrtb2.BidRequest) ([]*adapters.RequestData, []error) {
 	var requestData []*adapters.RequestData
-	networkAdunitMap := make(map[string][]adnAdunit)
+	networkAdunitMap := make(map[string][]adnRequestAdunit)
 	headers := setHeaders(ortbRequest)
 	var noCookies bool = false
 
 	for _, imp := range ortbRequest.Imp {
-		if imp.Banner == nil {
-			return nil, []error{&errortypes.BadInput{
-				Message: fmt.Sprintf("ignoring imp id=%s, Adnuntius supports only Banner", imp.ID),
-			}}
-		}
 
 		var bidderExt adapters.ExtImpBidder
 		if err := jsonutil.Unmarshal(imp.Ext, &bidderExt); err != nil {
@@ -218,7 +47,6 @@ func (a *adapter) generateRequests(ortbRequest openrtb2.BidRequest) ([]*adapters
 				Message: fmt.Sprintf("Error unmarshalling ExtImpBidder: %s", err.Error()),
 			}}
 		}
-
 		var adnuntiusExt openrtb_ext.ImpExtAdnunitus
 		if err := jsonutil.Unmarshal(bidderExt.Bidder, &adnuntiusExt); err != nil {
 			return nil, []error{&errortypes.BadInput{
@@ -235,17 +63,32 @@ func (a *adapter) generateRequests(ortbRequest openrtb2.BidRequest) ([]*adapters
 			network = adnuntiusExt.Network
 		}
 
-		adUnit := adnAdunit{
-			AuId:       adnuntiusExt.Auid,
-			TargetId:   fmt.Sprintf("%s-%s", adnuntiusExt.Auid, imp.ID),
-			Dimensions: getImpSizes(imp),
+		if imp.Banner != nil {
+			adUnit := generateAdUnit(imp, adnuntiusExt, "banner")
+			adUnit.AdType = ""
+
+			networkAdunitMap[network] = append(
+				networkAdunitMap[network],
+				adUnit)
 		}
-		if adnuntiusExt.MaxDeals > 0 {
-			adUnit.MaxDeals = adnuntiusExt.MaxDeals
+
+		if imp.Native != nil {
+			adUnit := generateAdUnit(imp, adnuntiusExt, "native")
+			adUnit.AdType = "NATIVE"
+			nativeRequest := json.RawMessage{}
+
+			if err := json.Unmarshal([]byte(imp.Native.Request), &nativeRequest); err != nil {
+				return nil, []error{&errortypes.BadInput{
+					Message: fmt.Sprintf("Error unmarshalling Native: %s", err.Error()),
+				}}
+			}
+
+			adUnit.NativeRequest.Ortb = nativeRequest
+			networkAdunitMap[network] = append(
+				networkAdunitMap[network],
+				adUnit)
 		}
-		networkAdunitMap[network] = append(
-			networkAdunitMap[network],
-			adUnit)
+
 	}
 
 	endpoint, err := makeEndpointUrl(ortbRequest, a, noCookies)
@@ -265,19 +108,19 @@ func (a *adapter) generateRequests(ortbRequest openrtb2.BidRequest) ([]*adapters
 		return nil, []error{fmt.Errorf("failed to parse site Ext: %v", err)}
 	}
 
+	var extUser openrtb_ext.ExtUser
+	if ortbRequest.User != nil && ortbRequest.User.Ext != nil {
+		if err := jsonutil.Unmarshal(ortbRequest.User.Ext, &extUser); err != nil {
+			return nil, []error{fmt.Errorf("failed to parse Ext User: %v", err)}
+		}
+	}
+
 	for _, networkAdunits := range networkAdunitMap {
 
 		adnuntiusRequest := adnRequest{
 			AdUnits:   networkAdunits,
 			Context:   site,
 			KeyValues: extSite.Data,
-		}
-
-		var extUser openrtb_ext.ExtUser
-		if ortbRequest.User != nil && ortbRequest.User.Ext != nil {
-			if err := jsonutil.Unmarshal(ortbRequest.User.Ext, &extUser); err != nil {
-				return nil, []error{fmt.Errorf("failed to parse Ext User: %v", err)}
-			}
 		}
 
 		// Will change when our adserver can accept multiple user IDS
@@ -315,6 +158,97 @@ func (a *adapter) generateRequests(ortbRequest openrtb2.BidRequest) ([]*adapters
 	return requestData, nil
 }
 
+func (a *adapter) MakeRequests(request *openrtb2.BidRequest, reqInfo *adapters.ExtraRequestInfo) ([]*adapters.RequestData, []error) {
+	return a.generateRequests(*request)
+}
+
+func generateBidResponse(adnResponse *AdnResponse, request *openrtb2.BidRequest) (*adapters.BidderResponse, []error) {
+	bidResponse := adapters.NewBidderResponseWithBidsCapacity(len(adnResponse.AdUnits))
+	var currency string
+	adunitMap := map[string]AdUnit{}
+	adunitMediaTypeMap := map[string][]AdUnit{}
+
+	/* Check the ad unit response to see if there are any multi ad  */
+	for _, adnRespAdunit := range adnResponse.AdUnits {
+		result := strings.Split(adnRespAdunit.TargetId, ":")
+		if adnRespAdunit.MatchedAdCount > 0 {
+			adunitMediaTypeMap[result[0]] = append(adunitMediaTypeMap[result[0]], adnRespAdunit)
+		}
+	}
+
+	/* Compare price if there are multiple media types */
+	for targetId, mappedAdunit := range adunitMediaTypeMap {
+		highestBidAtIndex := 0
+		if len(mappedAdunit) > 1 {
+			for index, mapped := range mappedAdunit {
+				fmt.Println(index, mapped.TargetId, mappedAdunit[index].Ads[0].Bid.Amount)
+				if mappedAdunit[index].Ads[0].Bid.Amount > mappedAdunit[highestBidAtIndex].Ads[0].Bid.Amount {
+					highestBidAtIndex = index
+				}
+			}
+		}
+		adunitMap[targetId] = mappedAdunit[highestBidAtIndex]
+	}
+
+	for _, imp := range request.Imp {
+
+		auId, _, _, err := jsonparser.Get(imp.Ext, "bidder", "auId")
+		if err != nil {
+			return nil, []error{&errortypes.BadInput{
+				Message: fmt.Sprintf("Error at Bidder auId: %s", err.Error()),
+			}}
+		}
+
+		targetID := fmt.Sprintf("%s-%s", string(auId), imp.ID)
+
+		adunit := adunitMap[targetID]
+
+		if len(adunit.Ads) > 0 {
+
+			ad := adunit.Ads[0]
+			html := adunit.Html
+			var bidType openrtb2.MarkupType = 1
+
+			currency = ad.Bid.Currency
+			native, _, _, _ := jsonparser.Get(adunit.NativeJson, "ortb")
+
+			if native != nil {
+				html = string(native)
+				bidType = 4
+			}
+
+			adBid, err := generateAdResponse(ad, imp, html, bidType, request)
+			if err != nil {
+				return nil, []error{&errortypes.BadInput{
+					Message: "Error at ad generation",
+				}}
+			}
+
+			bidResponse.Bids = append(bidResponse.Bids, &adapters.TypedBid{
+				Bid:     adBid,
+				BidType: convertMarkupTypeToBidType(bidType),
+			})
+
+			for _, deal := range adunit.Deals {
+				bidType = 1
+				dealBid, err := generateAdResponse(deal, imp, deal.Html, bidType, request)
+				if err != nil {
+					return nil, []error{&errortypes.BadInput{
+						Message: "Error at ad generation",
+					}}
+				}
+
+				bidResponse.Bids = append(bidResponse.Bids, &adapters.TypedBid{
+					Bid:     dealBid,
+					BidType: convertMarkupTypeToBidType(bidType),
+				})
+			}
+		}
+	}
+	bidResponse.Currency = currency
+	return bidResponse, nil
+}
+
 func (a *adapter) MakeBids(request *openrtb2.BidRequest, externalRequest *adapters.RequestData, response *adapters.ResponseData) (*adapters.BidderResponse, []error) {
 
 	if response.StatusCode == http.StatusBadRequest {
@@ -342,76 +276,7 @@ func (a *adapter) MakeBids(request *openrtb2.BidRequest, externalRequest *adapte
 	return bidResponse, nil
 }
 
-func getSiteExtAsKv(request *openrtb2.BidRequest) (siteExt, error) {
-	var extSite siteExt
-	if request.Site != nil && request.Site.Ext != nil {
-		if err := jsonutil.Unmarshal(request.Site.Ext, &extSite); err != nil {
-			return extSite, fmt.Errorf("failed to parse ExtSite in Adnuntius: %v", err)
-		}
-	}
-	return extSite, nil
-}
-
-func getGDPR(request *openrtb2.BidRequest) (string, string, error) {
-
-	gdpr := ""
-	var extRegs openrtb_ext.ExtRegs
-	if request.Regs != nil && request.Regs.Ext != nil {
-		if err := jsonutil.Unmarshal(request.Regs.Ext, &extRegs); err != nil {
-			return "", "", fmt.Errorf("failed to parse ExtRegs in Adnuntius GDPR check: %v", err)
-		}
-		if extRegs.GDPR != nil && (*extRegs.GDPR == 0 || *extRegs.GDPR == 1) {
-			gdpr = strconv.Itoa(int(*extRegs.GDPR))
-		}
-	}
-
-	consent := ""
-	if request.User != nil && request.User.Ext != nil {
-		var extUser openrtb_ext.ExtUser
-		if err := jsonutil.Unmarshal(request.User.Ext, &extUser); err != nil {
-			return "", "", fmt.Errorf("failed to parse ExtUser in Adnuntius GDPR check: %v", err)
-		}
-		consent = extUser.Consent
-	}
-
-	return gdpr, consent, nil
-}
-
-func generateReturnExt(ad Ad, request *openrtb2.BidRequest) (json.RawMessage, error) {
-	// We always force the publisher to render
-	var adRender int8 = 0
-
-	var requestRegsExt *openrtb_ext.ExtRegs
-	if request.Regs != nil && request.Regs.Ext != nil {
-		if err := jsonutil.Unmarshal(request.Regs.Ext, &requestRegsExt); err != nil {
-
-			return nil, fmt.Errorf("Failed to parse Ext information in Adnuntius: %v", err)
-		}
-	}
-
-	if ad.Advertiser.Name != "" && requestRegsExt != nil && requestRegsExt.DSA != nil {
-		legalName := ad.Advertiser.Name
-		if ad.Advertiser.LegalName != "" {
-			legalName = ad.Advertiser.LegalName
-		}
-		ext := &openrtb_ext.ExtBid{
-			DSA: &openrtb_ext.ExtBidDSA{
-				AdRender: &adRender,
-				Paid:     legalName,
-				Behalf:   legalName,
-			},
-		}
-		returnExt, err := json.Marshal(ext)
-		if err != nil {
-			return nil, fmt.Errorf("Failed to parse Ext information in Adnuntius: %v", err)
-		}
-
-		return returnExt, nil
-	}
-	return nil, nil
-}
-
-func generateAdResponse(ad Ad, imp openrtb2.Imp, html string, request *openrtb2.BidRequest) (*openrtb2.Bid, []error) {
+func generateAdResponse(ad Ad, imp openrtb2.Imp, html string, mType openrtb2.MarkupType, request *openrtb2.BidRequest) (*openrtb2.Bid, []error) {
 
 	creativeWidth, widthErr := strconv.ParseInt(ad.CreativeWidth, 10, 64)
 	if widthErr != nil {
@@ -427,8 +292,6 @@ func generateAdResponse(ad Ad, imp openrtb2.Imp, html string, request *openrtb2.
 		}}
 	}
 
-	price := ad.Bid.Amount
-
 	var bidderExt adapters.ExtImpBidder
 	if err := jsonutil.Unmarshal(imp.Ext, &bidderExt); err != nil {
 		return nil, []error{&errortypes.BadInput{
@@ -443,6 +306,7 @@ func generateAdResponse(ad Ad, imp openrtb2.Imp, html string, request *openrtb2.
 		}}
 	}
 
+	price := ad.Bid.Amount
 	if adnuntiusExt.BidType != "" {
 		if strings.EqualFold(string(adnuntiusExt.BidType), "net") {
 			price = ad.NetBid.Amount
@@ -470,68 +334,10 @@ func generateAdResponse(ad Ad, imp openrtb2.Imp, html string, request *openrtb2.
 		CrID:    ad.CreativeId,
 		Price:   price * 1000,
 		AdM:     html,
+		MType:   mType,
 		ADomain: ad.AdvertiserDomains,
 		Ext:     extJson,
 	}
+
 	return &bid, nil
-
-}
-
-func generateBidResponse(adnResponse *AdnResponse, request *openrtb2.BidRequest) (*adapters.BidderResponse, []error) {
-	bidResponse := adapters.NewBidderResponseWithBidsCapacity(len(adnResponse.AdUnits))
-	var currency string
-	adunitMap := map[string]AdUnit{}
-
-	for _, adnRespAdunit := range adnResponse.AdUnits {
-		adunitMap[adnRespAdunit.TargetId] = adnRespAdunit
-	}
-
-	for _, imp := range request.Imp {
-
-		auId, _, _, err := jsonparser.Get(imp.Ext, "bidder", "auId")
-		if err != nil {
-			return nil, []error{&errortypes.BadInput{
-				Message: fmt.Sprintf("Error at Bidder auId: %s", err.Error()),
-			}}
-		}
-
-		targetID := fmt.Sprintf("%s-%s", string(auId), imp.ID)
-		adunit := adunitMap[targetID]
-
-		if len(adunit.Ads) > 0 {
-
-			ad := adunit.Ads[0]
-			currency = ad.Bid.Currency
-
-			adBid, err := generateAdResponse(ad, imp, adunit.Html, request)
-			if err != nil {
-				return nil, []error{&errortypes.BadInput{
-					Message: "Error at ad generation",
-				}}
-			}
-
-			bidResponse.Bids = append(bidResponse.Bids, &adapters.TypedBid{
-				Bid:     adBid,
-				BidType: "banner",
-			})
-
-			for _, deal := range adunit.Deals {
-				dealBid, err := generateAdResponse(deal, imp, deal.Html, request)
-				if err != nil {
-					return nil, []error{&errortypes.BadInput{
-						Message: "Error at ad generation",
-					}}
-				}
-
-				bidResponse.Bids = append(bidResponse.Bids, &adapters.TypedBid{
-					Bid:     dealBid,
-					BidType: "banner",
-				})
-			}
-
-		}
-
-	}
-	bidResponse.Currency = currency
-	return bidResponse, nil
 }
