--- adapters/adnuntius/adnuntiustest/supplemental/banner-nil-check.json
@@ -1,29 +1,0 @@
-{
-	"mockBidRequest": {
-		"id": "test-request-id",
-		"user": {
-			"id": "1kjh3429kjh295jkl"
-		},
-		"device":{
-			"ext":{
-				"noCookies" : true
-			}
-		},
-		"imp": [
-			{
-				"id": "test-imp-id",
-				"ext": {
-					"bidder": {
-						"auId": "123"
-					}
-				}
-			}
-		]
-	},
-	"expectedMakeRequestsErrors": [
-		{
-			"value": "ignoring imp id=test-imp-id, Adnuntius supports only Banner",
-			"comparison": "literal"
-		}
-	]
-}
