+++ adapters/adnuntius/adnuntiustest/supplemental/check-dealId.json
@@ -36,8 +36,9 @@
                   "adUnits": [
                       {
                           "auId": "123",
-                          "targetId": "123-test-imp-id",
-            "dimensions": [[300,250],[300,600]]
+                          "targetId": "123-test-imp-id:banner",
+                          "nativeRequest": {},
+                          "dimensions": [[300,250],[300,600]]
                       }
                   ],
                   "context": "prebid.org",
