+++ adapters/adnuntius/adnuntiustest/supplemental/check-dsa-advertiser-legalName.json
@@ -50,7 +50,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,250],[300,600]]
 						}
 					],
