+++ adapters/adnuntius/adnuntiustest/supplemental/check-noCookies-parameter.json
@@ -36,7 +36,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,250],[300,600]]
 						}
 					],
