+++ adapters/adnuntius/adnuntiustest/supplemental/check-noCookies.json
@@ -40,7 +40,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,250],[300,600]]
 						}
 					],
