+++ adapters/adnuntius/adnuntiustest/supplemental/empty-regs-ext.json
@@ -38,7 +38,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,250],[300,600]]
 						}
 					],
