--- adapters/adnuntius/adnuntiustest/supplemental/native-error.json
@@ -1,24 +1,0 @@
-{
-	"mockBidRequest": {
-		"id": "unsupported-native-request",
-		"imp": [
-			{
-				"id": "unsupported-native-imp",
-				"native": {
-					"ver": "1.1"
-				},
-				"ext": {
-					"bidder": {
-						"auId": "1"
-					}
-				}
-			}
-		]
-	},
-	"expectedMakeRequestsErrors": [
-		{
-			"value": "ignoring imp id=unsupported-native-imp, Adnuntius supports only Banner",
-			"comparison": "literal"
-		}
-	]
-}
