+++ adapters/adnuntius/adnuntiustest/supplemental/size-check.json
@@ -33,7 +33,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,600]]
 						}
 					],
