+++ adapters/adnuntius/adnuntiustest/supplemental/user-ext.json
@@ -44,7 +44,8 @@
 					"adUnits": [
 						{
 							"auId": "123",
-							"targetId": "123-test-imp-id",
+							"targetId": "123-test-imp-id:banner",
+							"nativeRequest": {},
 							"dimensions": [[300,250],[300,600]]
 						}
 					],
