+++ adapters/adnuntius/params_test.go
@@ -41,6 +41,7 @@ func TestInvalidParams(t *testing.T) {
 var validParams = []string{
 	`{"auId":"123"}`,
 	`{"auId":"123", "network":"test"}`,
+	`{"auId":"123", "network":"test", "nativeRequest":{}}`,
 }
 
 var invalidParams = []string{
